pyproject.toml
[project]
name = "unsafe-docs"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pyyaml>=6.0.1",
  "markdowngenerator>=1.0.0",
]

[project.scripts]
unsafe-docs = "tools.unsafe_docs.cli:main"


